Add pagination to views.py functions
* Add pagination to investments_by_city function. I had to create a
session value to store the search value. If "?page" is not in url, this
is the first time we visit the page so session value is set to empty
character. Else it gets "entry" value. Session is reseted each time,
"?page" is not in url so when we have not clic on next or previous links